Detect deadlocks
Previously, the scheduler returned whenever there were no pending
events. This could be because the main function had finished, but it
could also happen if the system was deadlocked.

Now, the scheduler raises an exception if it finishes while the main
thread is still running, which should make the problem clearer.