Handle multiple exceptions in switches
If multiple threads fail, report all the exceptions. OCaml doesn't
provide a way to add context exceptions, so we have to introduce a new
exception type for this.

Also, `Switch.check` now wraps the exception in `Cancelled`. This allows
a thread to distinguish e.g. an `End_of_file` in its own operations
from being cancelled due to another thread hitting `End_of_file`.

Finally, `Fibre.fork_ignore ~sw` now copes with `sw` being off at the
start. It continues execution in the main thread, but does not run the
child. This is similar to the behaviour when `sw` is turned off
immediately after forking.