Simplify suspend operation
This replaces the old `Await` and `Yield` effects with a single
`Suspend` effect. The new effect just asks the scheduler to run a
user-provided function in the scheduler's context.

This moves most of the logic to Eio, avoiding the need to duplicate it
in each scheduler implementation.

Based on a suggestion by KC Sivaramakrishnan.