socket: set SO_REUSEPORT option
This commit adds the option of setting SO_REUSEPORT socket option in
linux_uring backend.

For libuv/luv backend, it seems it is not easy to set it as yet. This
is related to another outstanding issue related to SO_REUSEADDR
option in libuv backend. Perhaps once luv releases luv-unix, we can
revisit this issue again. For now in-order to ensure the repo builds
I have followed a similar approach as SO_REUSEADDR, i.e. dummy
function `luv_reuse_port` analogous to `luv_resuse_addr`.

SO_REUSEPORT can be used to build scalable, high performance network
servers in linux version 3.9 and over. This is so because reuseport
is used to enable socket sharding on multiple threads or child
processes. Listening sockets with SO_REUSEPORT enabled are further
load balanced by the kernel; so we get a performant load balancer
for free.

References:
https://www.nginx.com/blog/socket-sharding-nginx-release-1-9-1/
https://stackoverflow.com/questions/14388706/how-do-so-reuseaddr-and-so-reuseport-differ

Signed-off-by: Bikal Lem <gbikal+git@gmail.com>