Add `Fibre.fork_on_accept` and `Net.accept`
`fork_on_accept` runs an accept function in a new switch. If it
succeeds, it passes the result to a handler function in a new fibre,
which adopts the switch.

This allows moving the complicated accept logic out of the individual
backends. It also allows adding the simpler `Net.accept` function for
the case where only a single connection needs to be accepted.

The behaviour of `Fork.sub ~sw ~on_error` has changed slightly:
`on_error` now runs in the context of the parent `sw`,
and is now skipped if and only if `sw` is cancelled.
Also, the `on_release` argument has been removed, since its only use can
now be done using `fork_on_accept`.