Eio_linux: cope with lack of fixed chunks
`alloc` is now `alloc_fixed_or_wait` to make it clear that it waits
until space is available. The new `alloc_fixed` immediately returns an
option instead.

`free` is now `free_fixed` and `with_chunk` takes a fallback function to
use if no chunks are available.

The fixed buffer is rarely any faster than using regular memory, and
there is a risk of deadlock due to a shortage of chunks with the
blocking API. If we try to allocate a fixed buffer and fail, we now just
log a warning and continue without one.