Account for stack differences in the socketpair test (issue #312)
Linux returns ECONNRESET while every other stack tested returns EOF on this
test.

I did not dig into the especification but Linux appears to be on the weird side.

In order to get a ECONNRESET, the write needs to hit the socket _after_ the
close, but our program does:

WRITE(A)
CLOSE(B)
READ(A)

Linux never gives "A" side a chance to see a proper EOF, while the others do:

Linux        -> ECONNRESET
FreeBSD 12   -> EOF
OpenBSD 7.2  -> EOF
macos 12.5.1 -> EOF

Until we can reliably guarantee the same return on all stacks, relax the test a
bit and accept EOF.