Mbox implementation
Mboxes is an asynchronous/reliable/mpsc/domain-safe/single-value message passing
mechanism.

The mailbox holds one value only, senders block if a value is present, while
readers block while a value is not present.

- asynchronous since senders don't block (no rendezvous).
- reliable since messages cannot be lost (contrary to FFQ for example).
- MPSC since you can go nuts.
- domain-safe since any domain can do any combinatino of read/write across
  domains.
- single-value since the buffer contains only one value.

This is not intended for high performance, it's more of a inter-domain control
message passing, but no one will arrest you if you use it in the hot path.

Readers and Writers take turn and they're bounded, there is no thundering storm
on any side.

This diff also includes a tests/racers directory, the idea is to add programs
that try to race some part of the library.

racers/race_mbox.ml attempts to sends brazilions of messages making sure the
program doesn't hang, that messages are not lost and things are dandy.
Main strategy is having random sleeps on certain points and count sent vs
received messages, it's naive but it works ! I ran it for ~8 hours.