Expose `File.stat` and improve `Path.load`
Some notes:
  - Use Unix.Largestat instead of Unix.stat
  - Add a portable stat for the user
  - Moved Path.Unix_perm into File.Unix_perm to avoid cyclical ref

A previous attempt of this diff tried to make proper use of Luv.File.Stat but
there were multiple issues: `mode` was not properly exported so I had to add a
discovery.ml; it only works for a subset of file types and we had to manually
convert timestamps to float; all doable but an insuferable amount of code to end
up with less functionality than just calling into Unix.Largestat.stat.

There are still some kinks:
  - Luv will always follow a link, so we can't really stat on a symbolic link,
    this is because we rely on realpath(3) which resolves the link.
  - We can't open named pipes in with luv atm, so we also can't stat it (#357).
  - I did test all filemodes, but didn't add all to the test since they are
    impractical (opening /dev/tty for Character_special...)

--

Path.load was doubling the buffer and trying to fit enough until it saw an EOF.
This would cause many unecessary large allocations and my poor 16GB ram laptop
couldn't even load a 5GB file without getting OOMed.

This diff makes load create an initial buffer of the size of the file and load
exactly that amount of bytes in it.