Simplify cancellation logic
Before, the fiber context held an atomic recording whether the operation
had succeeded or been cancelled, to ensure that only one of these things
happened. However, this doesn't need to be atomic for single-threaded
operations, and even for parallel operations, we might perfer to keep
the atomic state elsewhere.

Now, operations just set a regular function and it's up to the operation
to ensure it's atomic if needed.

This is needed to implement lock-free waiters. Otherwise we have this
problem:

1. Domain 1 starts waiting for something from another domain.
2. Domain 2 prepares to provide a value to it.
3. Domain 1 cancels the operation and continues.
4. Domain 1 starts a new operation with a new cancel function.
5. Domain 2 sees the operation isn't cancelled and sends the value.

This isn't a problem when using a lock because both domains take the
lock. Currently, the waiters module still uses a lock (but now
provides its own mutex), and so works the same as before.

Some other related improvements:

- In eio_linux, we no longer wait for the cancel response before continuing.

  Before, cancelling an io_uring request would wait for the result of
  the cancellation before the cancel operation returned. But this isn't
  very useful, makes cancelling lots of things very slow since you can
  only do one at a time, and doesn't work with the new cancellation
  system where you can't switch fibers while cancelling.

- The "Cancelling multiple jobs" test has been removed. It was testing
  what happens when a cancel function switches fibers, but this is no
  longer allowed.

- In eio_luv, we only report cancellation if the operation didn't succeed.

  This changes the signature of `Low_level.await_with_cancel` slightly,
  since it now needs to know whether the operation was successful or not.

- Moved the cancellation logic into the domain manager implementation.
  This was the only place where a fiber could have a cancel function
  while not suspended.