eio_linux: drain ring on exit
Normally, all operations should have finished by the time we exit
because we don't exit until all fibers have finished, and a fiber can't
finish while an operation is in progress. However, this is not the case
for cancellation operations, which may still be active in rare cases.
So drain any remaining CQEs at exit.

Other options here are to allow cancellation operations to block (which
they used to do, but that caused other problems), or to have the main
operation wait for its cancellation too (but that's tricky and affects
the fast-path).

Fixes #467.