Unify Eio_linux.FD and Eio_posix.Fd as Eio_unix.Fd
Now that eio_luv is gone, there is no need for different backends to
have different types for wrapped file descriptors. This means we can use
wrapped FDs in more places, avoiding races.

This also allows e.g. `Eio_unix.pipe` to return resources with an Fd.t
attached explicitly in the type. It also avoids duplicating the
fork actions in each backend.

In this commit the old APIs are mostly still present, and the tests
still use them.