Adopt OCaml Code of Conduct
This code of conduct lives in [ocaml/code-of-conduct](https://github.com/ocaml/code-of-conduct) and has been discussed in [this thread](https://discuss.ocaml.org/t/ocaml-community-code-of-conduct/10494).

It has been adopted by the OCaml compiler in [ocaml/ocaml#11761](https://github.com/ocaml/ocaml/pull/11761) and many platform tools. After a discussion with maintainers, we propose adopting it for Eio.